Conflicts:
	KeywordSearch/src/org/sleuthkit/autopsy/keywordsearch/Bundle.properties
	KeywordSearch/src/org/sleuthkit/autopsy/keywordsearch/docs/keywordsearch-about.html
	test/script/regression.py
	test/script/srcupdater.py

+++ RecentActivity/src/org/sleuthkit/autopsy/recentactivity/Firefox.java
@@ -56,7 +56,8 @@ public class Firefox extends Extract {
     private static final String cookieQuery = "SELECT name,value,host,expiry,(lastAccessed/1000000) as lastAccessed,(creationTime/1000000) as creationTime FROM moz_cookies";
     private static final String cookieQueryV3 = "SELECT name,value,host,expiry,(lastAccessed/1000000) as lastAccessed FROM moz_cookies";
     private static final String bookmarkQuery = "SELECT fk, moz_bookmarks.title, url, (moz_bookmarks.dateAdded/1000000) as dateAdded FROM moz_bookmarks INNER JOIN moz_places ON moz_bookmarks.fk=moz_places.id";
-    private static final String downloadQuery = "SELECT target, source,(startTime/1000000) as startTime, maxBytes  FROM moz_downloads";   
+    private static final String downloadQuery = "SELECT target, source,(startTime/1000000) as startTime, maxBytes  FROM moz_downloads";
+    private static final String downloadQueryVersion24 = "SELECT url, content as target, (lastModified/1000000) as lastModified FROM moz_places, moz_annos WHERE moz_places.id = moz_annos.place_id AND moz_annos.anno_attribute_id = 3";
     
     public int FireFoxCount = 0;
     final public static String MODULE_VERSION = "1.0";
@@ -73,7 +74,8 @@ public class Firefox extends Extract {
     }
 
     @Override
-    public void process(PipelineContext<IngestModuleDataSource>pipelineContext, Content dataSource, IngestDataSourceWorkerController controller) {
+    public void process(PipelineContext<IngestModuleDataSource> pipelineContext, Content dataSource, IngestDataSourceWorkerController controller) {
+        dataFound = false;
         this.getHistory(dataSource, controller);
         this.getBookmark(dataSource, controller);
         this.getDownload(dataSource, controller);
@@ -93,8 +95,17 @@ public class Firefox extends Extract {
             String msg = "Error fetching internet history files for Firefox.";
             logger.log(Level.WARNING, msg);
             this.addErrorMessage(this.getName() + ": " + msg);
+            return;
+        }
+        
+        if (historyFiles.isEmpty()) {
+            String msg = "No FireFox history files found.";
+            logger.log(Level.INFO, msg);
+            return;
         }
 
+        dataFound = true;
+        
         int j = 0;
         for (AbstractFile historyFile : historyFiles) {
             if (historyFile.getSize() == 0) {
@@ -121,11 +132,9 @@ public class Firefox extends Extract {
                 Collection<BlackboardAttribute> bbattributes = new ArrayList<BlackboardAttribute>();
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL.getTypeID(), "RecentActivity", ((result.get("url").toString() != null) ? result.get("url").toString() : "")));
                 //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL_DECODED.getTypeID(), "RecentActivity", ((result.get("url").toString() != null) ? EscapeUtil.decodeURL(result.get("url").toString()) : "")));
-                //TODO Revisit usage of deprecated constructor as per TSK-583
-                //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_LAST_ACCESSED.getTypeID(), "RecentActivity", "Last Visited", (Long.valueOf(result.get("visit_date").toString()))));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DATETIME_ACCESSED.getTypeID(), "RecentActivity", (Long.valueOf(result.get("visit_date").toString()))));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_REFERRER.getTypeID(), "RecentActivity", ((result.get("ref").toString() != null) ? result.get("ref").toString() : "")));
-                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_NAME.getTypeID(), "RecentActivity", ((result.get("title").toString() != null) ? result.get("title").toString() : "")));
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_TITLE.getTypeID(), "RecentActivity", ((result.get("title").toString() != null) ? result.get("title").toString() : "")));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PROG_NAME.getTypeID(), "RecentActivity", "FireFox"));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DOMAIN.getTypeID(), "RecentActivity", (Util.extractDomain((result.get("url").toString() != null) ? result.get("url").toString() : ""))));
                 this.addArtifact(ARTIFACT_TYPE.TSK_WEB_HISTORY, historyFile, bbattributes);
@@ -156,6 +165,13 @@ public class Firefox extends Extract {
             return;
         }
 
+        if (bookmarkFiles.isEmpty()) {
+            logger.log(Level.INFO, "Didn't find any firefox bookmark files.");
+            return;
+        }
+        
+        dataFound = true;
+        
         int j = 0;
         for (AbstractFile bookmarkFile : bookmarkFiles) {
             if (bookmarkFile.getSize() == 0) {
@@ -181,8 +197,10 @@ public class Firefox extends Extract {
 
                 Collection<BlackboardAttribute> bbattributes = new ArrayList<BlackboardAttribute>();
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL.getTypeID(), "RecentActivity", ((result.get("url").toString() != null) ? result.get("url").toString() : "")));
-                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_NAME.getTypeID(), "RecentActivity", ((result.get("title").toString() != null) ? result.get("title").toString() : "")));
-                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DATETIME_ACCESSED.getTypeID(), "RecentActivity", (Long.valueOf(result.get("dateAdded").toString()))));
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_TITLE.getTypeID(), "RecentActivity", ((result.get("title").toString() != null) ? result.get("title").toString() : "")));
+                if (Long.valueOf(result.get("dateAdded").toString()) > 0) {
+                    bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DATETIME_CREATED.getTypeID(), "RecentActivity", (Long.valueOf(result.get("dateAdded").toString()))));
+                }
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PROG_NAME.getTypeID(), "RecentActivity", "FireFox"));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DOMAIN.getTypeID(), "RecentActivity", (Util.extractDomain((result.get("url").toString() != null) ? result.get("url").toString() : ""))));
                 this.addArtifact(ARTIFACT_TYPE.TSK_WEB_BOOKMARK, bookmarkFile, bbattributes);
@@ -212,6 +230,12 @@ public class Firefox extends Extract {
             return;
         }
 
+        if (cookiesFiles.isEmpty()) {
+            logger.log(Level.INFO, "Didn't find any Firefox cookie files.");
+            return;
+        }
+        
+        dataFound = true;
         int j = 0;
         for (AbstractFile cookiesFile : cookiesFiles) {
             if (cookiesFile.getSize() == 0) {
@@ -264,14 +288,27 @@ public class Firefox extends Extract {
 
         services.fireModuleDataEvent(new ModuleDataEvent("Recent Activity", BlackboardArtifact.ARTIFACT_TYPE.TSK_WEB_COOKIE));
     }
-
+    
     /**
      * Queries for downloads files and adds artifacts
      * @param dataSource
      * @param controller 
      */
     private void getDownload(Content dataSource, IngestDataSourceWorkerController controller) {
+        getDownloadPreVersion24(dataSource, controller);
+        getDownloadVersion24(dataSource, controller);
+    }
 
+    /**
+     * Finds downloads artifacts from Firefox data from versions before 24.0.
+     * 
+     * Downloads were stored in a separate downloads database.
+     * 
+     * @param dataSource
+     * @param controller 
+     */
+    private void getDownloadPreVersion24(Content dataSource, IngestDataSourceWorkerController controller) {
+        
         FileManager fileManager = currentCase.getServices().getFileManager();
         List<AbstractFile> downloadsFiles = null;
         try {
@@ -282,7 +319,13 @@ public class Firefox extends Extract {
             this.addErrorMessage(this.getName() + ": " + msg);
             return;
         }
-
+        
+        if (downloadsFiles.isEmpty()) {
+            logger.log(Level.INFO, "Didn't find any pre-version-24.0 Firefox download files.");
+            return;
+        }
+        
+        dataFound = true;
         int j = 0;
         for (AbstractFile downloadsFile : downloadsFiles) {
             if (downloadsFile.getSize() == 0) {
@@ -312,19 +355,24 @@ public class Firefox extends Extract {
 
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL.getTypeID(), "RecentActivity", ((result.get("source").toString() != null) ? result.get("source").toString() : "")));
                 //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL_DECODED.getTypeID(), "RecentActivity", ((result.get("source").toString() != null) ? EscapeUtil.decodeURL(result.get("source").toString()) : "")));
-                //TODO Revisit usage of deprecated constructor as per TSK-583
-                //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_LAST_ACCESSED.getTypeID(), "RecentActivity", "Last Visited", (Long.valueOf(result.get("startTime").toString()))));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DATETIME_ACCESSED.getTypeID(), "RecentActivity", (Long.valueOf(result.get("startTime").toString()))));
                 
-                try {
-                    String urldecodedtarget = URLDecoder.decode(result.get("source").toString().replaceAll("file:///", ""), "UTF-8");
-                    bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH_ID.getTypeID(), "RecentActivity", Util.findID(dataSource, urldecodedtarget)));
-                } catch (UnsupportedEncodingException ex) {
-                    logger.log(Level.SEVERE, "Error decoding Firefox download URL in " + temps, ex);
-                    errors++;
+                String target = result.get("target").toString();
+                
+                if (target != null) {
+                    try {
+                        String decodedTarget = URLDecoder.decode(target.toString().replaceAll("file:///", ""), "UTF-8");
+                        bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH.getTypeID(), "RecentActivity", decodedTarget));
+                        long pathID = Util.findID(dataSource, decodedTarget);
+                        if (pathID != -1) {
+                            bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH_ID.getTypeID(), "RecentActivity", pathID));
+                        }
+                    } catch (UnsupportedEncodingException ex) {
+                        logger.log(Level.SEVERE, "Error decoding Firefox download URL in " + temps, ex);
+                        errors++;
+                    }
                 }
                     
-                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH.getTypeID(), "RecentActivity", ((result.get("target").toString() != null) ? result.get("target").toString() : "")));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PROG_NAME.getTypeID(), "RecentActivity", "FireFox"));
                 bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DOMAIN.getTypeID(), "RecentActivity", (Util.extractDomain((result.get("source").toString() != null) ? result.get("source").toString() : ""))));
                 this.addArtifact(ARTIFACT_TYPE.TSK_WEB_DOWNLOAD, downloadsFile, bbattributes);
@@ -332,6 +380,96 @@ public class Firefox extends Extract {
             }
             if (errors > 0) {
                 this.addErrorMessage(this.getName() + ": Error parsing " + errors + " Firefox web history artifacts.");
+            }
+            j++;
+            dbFile.delete();
+            break;
+        }
+        
+        services.fireModuleDataEvent(new ModuleDataEvent("Recent Activity", BlackboardArtifact.ARTIFACT_TYPE.TSK_WEB_DOWNLOAD));
+    }
+    
+    /**
+     * Gets download artifacts from Firefox data from version 24.
+     * 
+     * Downloads are stored in the places database.
+     * 
+     * @param dataSource
+     * @param controller 
+     */
+    private void getDownloadVersion24(Content dataSource, IngestDataSourceWorkerController controller) {
+        FileManager fileManager = currentCase.getServices().getFileManager();
+        List<AbstractFile> downloadsFiles = null;
+        try {
+            downloadsFiles = fileManager.findFiles(dataSource, "places.sqlite", "Firefox");
+        } catch (TskCoreException ex) {
+            String msg = "Error fetching 'downloads' files for Firefox.";
+            logger.log(Level.WARNING, msg);
+            this.addErrorMessage(this.getName() + ": " + msg);
+            return;
+        }
+        
+        if (downloadsFiles.isEmpty()) {
+            logger.log(Level.INFO, "Didn't find any version-24.0 Firefox download files.");
+            return;
+        }
+        
+        dataFound = true;
+        int j = 0;
+        for (AbstractFile downloadsFile : downloadsFiles) {
+            if (downloadsFile.getSize() == 0) {
+                continue;
+            }
+            String fileName = downloadsFile.getName();
+            String temps = RAImageIngestModule.getRATempPath(currentCase, "firefox") + File.separator + fileName + "-downloads" + j + ".db";
+            int errors = 0;
+            try {
+                ContentUtils.writeToFile(downloadsFile, new File(temps));
+            } catch (IOException ex) {
+                logger.log(Level.SEVERE, "Error writing the sqlite db for firefox download artifacts.{0}", ex);
+                this.addErrorMessage(this.getName() + ": Error while trying to analyze file:" + fileName);
+                continue;
+            }
+            File dbFile = new File(temps);
+            if (controller.isCancelled()) {
+                dbFile.delete();
+                break;
+            }
+            
+            List<HashMap<String, Object>> tempList = this.dbConnect(temps, downloadQueryVersion24);
+            
+            logger.log(Level.INFO, moduleName + "- Now getting downloads from " + temps + " with " + tempList.size() + "artifacts identified.");
+            for (HashMap<String, Object> result : tempList) {
+                
+                Collection<BlackboardAttribute> bbattributes = new ArrayList<BlackboardAttribute>();
+
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL.getTypeID(), "RecentActivity", ((result.get("url").toString() != null) ? result.get("url").toString() : "")));
+                //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_URL_DECODED.getTypeID(), "RecentActivity", ((result.get("source").toString() != null) ? EscapeUtil.decodeURL(result.get("source").toString()) : "")));
+                //TODO Revisit usage of deprecated constructor as per TSK-583
+                //bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_LAST_ACCESSED.getTypeID(), "RecentActivity", "Last Visited", (Long.valueOf(result.get("startTime").toString()))));
+                
+                String target = result.get("target").toString();
+                if (target != null) {
+                    try {
+                        String decodedTarget = URLDecoder.decode(target.toString().replaceAll("file:///", ""), "UTF-8");
+                        bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH.getTypeID(), "RecentActivity", decodedTarget));
+                        long pathID = Util.findID(dataSource, decodedTarget);
+                        if (pathID != -1) {
+                            bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PATH_ID.getTypeID(), "RecentActivity", pathID));
+                        }
+                    } catch (UnsupportedEncodingException ex) {
+                        logger.log(Level.SEVERE, "Error decoding Firefox download URL in " + temps, ex);
+                        errors++;
+                    }
+                }
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DATETIME_ACCESSED.getTypeID(), "RecentActivity", Long.valueOf(result.get("lastModified").toString())));
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_PROG_NAME.getTypeID(), "RecentActivity", "FireFox"));
+                bbattributes.add(new BlackboardAttribute(ATTRIBUTE_TYPE.TSK_DOMAIN.getTypeID(), "RecentActivity", (Util.extractDomain((result.get("url").toString() != null) ? result.get("url").toString() : ""))));
+                this.addArtifact(ARTIFACT_TYPE.TSK_WEB_DOWNLOAD, downloadsFile, bbattributes);
+                
+            }
+            if (errors > 0) {
+                this.addErrorMessage(this.getName() + ": Error parsing " + errors + " Firefox web download artifacts.");
             }
             j++;
             dbFile.delete();
